Original bot code (from MCBot made Fall 2018)
MirandaCosgroveBot has died a noble death, DingusBot will replace her with a lot of new features and optimization fixes. Since I have started working a lot on this project on my own after the dissolution of my team I decided to rename the bot so that I can preserve the (except a few changes I made after the semester) finished product of the Miranda-Cosgroves team.